mlir: add support for graph level code generation
Add generation of mlir operations inlined and allocation
of temporary tensors through Memref Alloca operations.

Update scheduler interface to take nodes parameters as a list
of nodes names to schedule.

Update node naming for MlirOps from *_fill to *_0 and *_reduce to *.
Hence, the default is to have the last node name with the same name as the
operation name. The *_0 convention is to say input 0 of the last node.